Another type error, tried to divide an int to produce a range between 0 and 255

+++ lab02/main.c
@@ -51,7 +51,7 @@ int main (int argc, char *argv[]) {
 	}
 	*/
 
-	float brightness_pct = brightness/255; // between 0 and 1
+	float brightness_pct = (float) brightness/255; // between 0 and 1
 	printf("Your desired brightness: %f\n",brightness_pct);
 	
 	int i;
